Files and Tests
Added one test for the hamiltonian function, moved it to a file of it's own. Fixed a sign issue.
